1078-解析104...

+++ src/main/java/com/genersoft/iot/vmp/jt1078/bean/common/ConfigAttribute.java
@@ -6,7 +6,7 @@ import java.lang.annotation.RetentionPolicy;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface ConfigAttribute {
 
-    byte id();
+    long id();
 
     String description();
 }
